\name{BuildHeader}
\alias{BuildHeader}
\title{BuildHeader}
\usage{
BuildHeader()
}
\value{
Header string for the old API auth method.
}
\description{
Internal function - Build Header for REST API call. This is
deprecated, we should now be using OAUTH.
}
\seealso{
Other internal: \code{\link{ApiRequest}};
\code{\link{BuildInnerBreakdownsRecursively}};
\code{\link{ParseFallout}}; \code{\link{ParseOvertime}};
\code{\link{ParsePathing}}; \code{\link{ParseRanked}};
\code{\link{ParseTrended}}
}

