% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInnerBreakdownsRecursively.R
\name{BuildInnerBreakdownsRecursively}
\alias{BuildInnerBreakdownsRecursively}
\title{BuildInnerBreakdownsRecursively}
\usage{
BuildInnerBreakdownsRecursively(parent.element, elements, metrics,
  current.recursion.level, context, accumulator = data.frame(),
  date.range = "")
}
\arguments{
\item{parent.element}{element containing breakdown rows}

\item{elements}{list of all elements included in the report}

\item{metrics}{list of all metrics included in the report}

\item{current.recursion.level}{current recursion level, initial call should set this to 0, this increments as the function is called recursively}

\item{context}{initially a blank list, with additional context added as we go further down the data structure}

\item{accumulator}{data frame used to accumulate the report data 
(yes, I use some for loops here, if anyone can find a better way to do this with hierarchical data, please submit a pull request)}

\item{date.range}{date range for the report, these columns are added if specified}
}
\value{
Flat data frame containing all key report data
}
\description{
Internal function - Builds inner breakdowns from nested report structure
}
\keyword{internal}
