% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BuildRealTimeReportStructure}
\alias{BuildRealTimeReportStructure}
\title{Build Configuration for Real-Time Report}
\usage{
BuildRealTimeReportStructure(report.name = "", metric = "",
  elements = c(), min.granularity = "1", ui.report = TRUE)
}
\arguments{
\item{report.name}{Real-Time Report Name}

\item{metric}{Metric for Real-Time Report}

\item{elements}{Breakdowns for Real-Time Report}

\item{min.granularity}{min.granularity}

\item{ui.report}{Show report in Adobe Analytics interface}
}
\value{
List
}
\description{
Selects the metrics and elements (dimensions) on which you want real time
reports enabled. Use individual reports as arguments in SaveRealTimeSettings
}
\examples{
\dontrun{

   }
}
\seealso{
\code{\link{GetRealTimeSettings}} \cr

\code{\link{SaveRealTimeSettings}} \cr
}
\keyword{BuildRealTimeReportStructure}

