\name{CancelReport}
\alias{CancelReport}
\title{Cancel a Report in the Report Queue}
\usage{
CancelReport(report.id)
}
\arguments{
  \item{report.id}{id of the report that you want to
  cancel}
}
\value{
Console message
}
\description{
CancelReport
}
\details{
Cancels the specified report
}
\examples{
\dontrun{
CancelReport('12345678')
}
}

