% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_simple_listcol.R
\name{collapse_simple_listcol}
\alias{collapse_simple_listcol}
\title{collapse_simple_listcol}
\usage{
collapse_simple_listcol(x)
}
\arguments{
\item{x}{An unnested list of length 1}
}
\value{
A collapsed atomic vector, with each element of \code{x[[1]]}
separated by \code{, } (comma\code{space}). If a comma is encountered
in any element of \code{x[[1]]}, ALL elements will additionally be 
encapsulated by square brackets as a more readable proxy for 
quotes.
}
\description{
Internal function - unnest a list-column into a (possibly bracket-quoted) comma-delimited atomic vector
}
\details{
This does not check for any upsteam (e.g. data.frame) names, and should be used
directly on a column accessed by name (or double-subscripted via index/name); mainly
a helper function that does the heavy lifting for collapse_tags and collapse_compatibility

See \code{\link{collapse_simple_target}}, where this function does the heavy lifting.
}
\examples{
# TBD
}
\seealso{
Other internal: \code{\link{GetEndpoint}},
  \code{\link{GetUsageLog}}, \code{\link{ParseDW}},
  \code{\link{ParseFallout}}, \code{\link{ParseOvertime}},
  \code{\link{ParsePathing}}, \code{\link{ParseRanked}},
  \code{\link{ParseSummary}}, \code{\link{ParseTrended}},
  \code{\link{call.Get_base}},
  \code{\link{collapse_simple_target}},
  \code{\link{split_segment_ret}}
}
