% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetActivation}
\alias{GetActivation}
\title{Get Activation Detail for a Report Suite(s)}
\usage{
GetActivation(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get activation detail for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
active <- GetActivation("your_report_suite")

active2 <- GetActivation(report_suites$rsid)
}
}

