% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAxleStartDate.R
\name{GetAxleStartDate}
\alias{GetAxleStartDate}
\title{Get Cutover Date from SC14 to SC15 for a Report Suite(s)}
\usage{
GetAxleStartDate(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get cutover date from SC14 to SC15 for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
switch <- GetAxleStartDate("your_report_suite")

switch2 <- GetAxleStartDate(report_suites$rsid)
}
}
