% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBaseURL.R
\name{GetBaseURL}
\alias{GetBaseURL}
\title{Get Base URL for a Report Suite(s)}
\usage{
GetBaseURL(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get base url for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
url <- GetBaseURL("your_report_suite")

url2 <- GetBaseURL(report_suites$rsid)
}
}
