% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetBookmarks}
\alias{GetBookmarks}
\title{Get Defined Bookmarks for a user}
\usage{
GetBookmarks(folder.limit = "", folder.offset = "")
}
\arguments{
\item{folder.limit}{}

\item{folder.offset}{}
}
\value{
Data frame
}
\description{
Get defined bookmarks for a user.
}
\details{
This function's arguments are both optional
}
\examples{
\dontrun{
bookmarks<- GetBookmarks()

bookmarks2 <- GetBookmarks('5', '1')
}
}

