% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCalculatedMetrics.R
\name{GetCalculatedMetrics}
\alias{GetCalculatedMetrics}
\title{Get Calculated Metrics for a Report Suite(s)}
\usage{
GetCalculatedMetrics(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get calculated metrics for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
calc<- GetCalculatedMetrics("your_report_suite")

calc2 <- GetCalculatedMetrics(report_suites$rsid)
}
}
