% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClickMapReporting.R
\name{GetClickMapReporting}
\alias{GetClickMapReporting}
\title{Get Click Map Settings for a Report Suite(s)}
\usage{
GetClickMapReporting(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get Click Map Settings for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
cmsettings<- GetClickMapReporting("your_report_suite")

cmsettings2 <- GetClickMapReporting(report_suites$rsid)
}
}
