% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetDashboards}
\alias{GetDashboards}
\title{Get Defined Dashboards}
\usage{
GetDashboards(dashboard.limit = "", dashboard.offset = "")
}
\arguments{
\item{dashboard.limit}{}

\item{dashboard.offset}{}
}
\value{
List
}
\description{
Get defined dashboards
}
\details{
This function's arguments are both optional
}
\examples{
\dontrun{
dash<- GetDashboards()

dash2 <- GetBookmarks('5', '1')
}
}

