% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefaultPage.R
\name{GetDefaultPage}
\alias{GetDefaultPage}
\title{Get Default Page for a Report Suite(s)}
\usage{
GetDefaultPage(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get default page for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
defpage <- GetDefaultPage("your_report_suite")

defpage2 <- GetDefaultPage(report_suites$rsid)
}
}
