% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDiscoverEnabled.R
\name{GetDiscoverEnabled}
\alias{GetDiscoverEnabled}
\title{Get Whether Discover is Enabled for a Report Suite(s)}
\usage{
GetDiscoverEnabled(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get whether Discover is enabled for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
discenable <- GetDiscoverEnabled("your_report_suite")

discenable2 <- GetDiscoverEnabled(report_suites$rsid)
}
}
