% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEndpoint.R
\name{GetEndpoint}
\alias{GetEndpoint}
\title{GetEndpoint}
\usage{
GetEndpoint(company)
}
\arguments{
\item{company}{Adobe Analytics Report Suite Company}
}
\value{
String containing API endpoint
}
\description{
Internal function that gets proper API endpoint for a given 
Adobe Analytics Company
}
\seealso{
Other internal: \code{\link{GetUsageLog}},
  \code{\link{ParseDW}}, \code{\link{ParseFallout}},
  \code{\link{ParseOvertime}}, \code{\link{ParsePathing}},
  \code{\link{ParseRanked}}, \code{\link{ParseSummary}},
  \code{\link{ParseTrended}}, \code{\link{call.Get_base}},
  \code{\link{collapse_simple_listcol}},
  \code{\link{collapse_simple_target}},
  \code{\link{split_segment_ret}}
}
\keyword{internal}
