% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{GetFunctions}
\alias{GetFunctions}
\title{Get Functions Defined in Adobe Analytics}
\usage{
GetFunctions()
}
\value{
Data Frame
}
\description{
Requests the id and definitions of functions in Adobe Analytics.
}
\details{
Returns descriptions/formulas available within Adobe Analytics such as median, pi, regression, etc.
}
\examples{
\dontrun{
aa_functions <- GetFunctions()
}
}
