% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeoSegmentation.R
\name{GetGeoSegmentation}
\alias{GetGeoSegmentation}
\title{Get the Geography Segmentation for a Report Suite(s)}
\usage{
GetGeoSegmentation(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get the geography segmentation for the requested report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
geoseg <- GetGeoSegmentation("your_report_suite")

geoseg2 <- GetGeoSegmentation(report_suites$rsid)
}
}
