% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGroup.R
\name{GetGroup}
\alias{GetGroup}
\title{Retrieves Information About The Specified Permission Group.}
\usage{
GetGroup(group_name)
}
\arguments{
\item{group_name}{The name of the group that you want to retrieve.}
}
\value{
Data frame
}
\description{
Retrieves information about the specified permission group.
}
\details{
This function requires having a group name specified.
}
\examples{
\dontrun{
group <- GetGroup("group_name")


}
}
