% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGroups.R
\name{GetGroups}
\alias{GetGroups}
\title{Get Defined User Groups for a Company}
\usage{
GetGroups(group_search_field = "", group_search_value = "")
}
\arguments{
\item{group_search_field}{Optional. Field to search for a specific value}

\item{group_search_value}{Optional. Use with group_search_field to search for a specific value}
}
\value{
Data frame
}
\description{
Get defined user groups for a company.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
groups <- GetGroups()


}
}
