% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInternalURLFilters.R
\name{GetInternalURLFilters}
\alias{GetInternalURLFilters}
\title{Get Internal URL Filters for a Report Suite(s)}
\usage{
GetInternalURLFilters(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get internal url fitlers for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
calc<- GetInternalURLFilters("your_report_suite")

calc2 <- GetInternalURLFilters(report_suites$rsid)
}
}
