% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetIPAddressExclusions}
\alias{GetIPAddressExclusions}
\title{Get the IP Address Exclusions for a Report Suite(s)}
\usage{
GetIPAddressExclusions(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get the IP address exclusions for the requested report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
ipexc <- GetIPAddressExclusions("your_report_suite")

ipexc2 <- GetIPAddressExclusions(report_suites$rsid)
}
}

