% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIPObfuscation.R
\name{GetIPObfuscation}
\alias{GetIPObfuscation}
\title{Get IP Obfuscation Status for a Report Suite(s)}
\usage{
GetIPObfuscation(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get IP Obfuscation status for the requested report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
ipobf <- GetIPObfuscation("your_report_suite")

ipobf2 <- GetIPObfuscation(report_suites$rsid)
}
}
