% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKeyVisitors.R
\name{GetKeyVisitors}
\alias{GetKeyVisitors}
\title{Get Key Visitors for a Report Suite(s)}
\usage{
GetKeyVisitors(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get key visitors for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
keyvisit<- GetKeyVisitors("your_report_suite")

keyvisit2 <- GetKeyVisitors(report_suites$rsid)
}
}
