% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetListVariables.R
\name{GetListVariables}
\alias{GetListVariables}
\title{Get List Variables for a Report Suite(s)}
\usage{
GetListVariables(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get list variables for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
calc<- GetListVariables("your_report_suite")

calc2 <- GetListVariables(report_suites$rsid)
}
}
