% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLocalization.R
\name{GetLocalization}
\alias{GetLocalization}
\title{Get Localization for a Report Suite(s)}
\usage{
GetLocalization(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get localization for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
local <- GetLocalization("your_report_suite")

local2 <- GetLocalization(report_suites$rsid)
}
}
