% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogin.R
\name{GetLogin}
\alias{GetLogin}
\title{Get Login Information for a Single Login}
\usage{
GetLogin(login)
}
\arguments{
\item{login}{Login to get information about}
}
\value{
List
}
\description{
Get login information for a single login
}
\details{
This function requires a character string as the input
}
\examples{
\dontrun{
login_info<- GetLogin('login-name')

}
}
