% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogins.R
\name{GetLogins}
\alias{GetLogins}
\title{Get Logins for a Company}
\usage{
GetLogins(login.search.field = "", login.search.value = "")
}
\arguments{
\item{login.search.field}{Search login, first_name, last_name or title}

\item{login.search.value}{Value to search for (case-insensitive)}
}
\value{
Data frame
}
\description{
Get logins for a company
}
\details{
This function's arguments are both optional
}
\examples{
\dontrun{
logins<- GetLogins()

logins2 <- GetLogins('last_name', 'zwitch')
}
}
