% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarketingChannelRules.R
\name{GetMarketingChannelRules}
\alias{GetMarketingChannelRules}
\title{Get Marketing Channel Rules for a Report Suite(s)}
\usage{
GetMarketingChannelRules(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get marketing channel rules for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
expire <- GetMarketingChannelRules("your_report_suite")

expire2 <- GetMarketingChannelRules(report_suites$rsid)
}
}
