% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarketingChannels.R
\name{GetMarketingChannels}
\alias{GetMarketingChannels}
\title{Get Defined Marketing Channels for a Report Suite(s)}
\usage{
GetMarketingChannels(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get defined marketing channels for each of the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
mchan <- GetMarketingChannels("your_report_suite")

mchan2 <- GetMarketingChannels(report_suites$rsid)
}
}
