% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMobileAppReporting.R
\name{GetMobileAppReporting}
\alias{GetMobileAppReporting}
\title{Get Mobile App Reporting Status for a Report Suite(s)}
\usage{
GetMobileAppReporting(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get mobile app reporting status for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
mobile <- GetMobileAppReporting("your_report_suite")

mobile2 <- GetMobileAppReporting(report_suites$rsid)
}
}
