% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPermanentTraffic.R
\name{GetPermanentTraffic}
\alias{GetPermanentTraffic}
\title{Get Permanent Traffic Setting for a Report Suite(s)}
\usage{
GetPermanentTraffic(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get permanent traffic setting for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
permtraf <- GetPermanentTraffic("your_report_suite")

permtraf2 <- GetPermanentTraffic(report_suites$rsid)
}
}
