% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrivacySettings.R
\name{GetPrivacySettings}
\alias{GetPrivacySettings}
\title{Get Privacy Settings for a Report Suite(s)}
\usage{
GetPrivacySettings(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get privacy settings for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{

privacy <- GetPrivacySettings("your_report_suite")

privacy2 <- GetPrivacySettings(c("your_dev_suite", "your_prod_suite"))

}
}
