% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReport.R
\name{GetReport}
\alias{GetReport}
\title{Get EnQueued Report by report ID}
\usage{
GetReport(report.id, interval.seconds = 10, max.attempts = 3,
  print.attempts = TRUE, format = "json", page = 0)
}
\arguments{
\item{report.id}{report id that's returned by QueueTrended and other functions while used with enqueueOnly parameter set to TRUE}

\item{interval.seconds}{How long to wait between attempts}

\item{max.attempts}{Number of API attempts before stopping}

\item{print.attempts}{Print each attempt for fetching data}

\item{format}{"csv" or "json"}

\item{page}{Page Number of Resuts (QueueDataWarehouse only)}
}
\value{
Data frame
}
\description{
Get a single report by report id, this allow asynchronous way of getting reports.
}
\details{
This is a function for advanced users, after you've enqueued multiple reports
and want to get one of them when it's ready.
}
\examples{
\dontrun{

custom_report <- GetReport(12345678)

}

}
