% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetReportDescription}
\alias{GetReportDescription}
\title{Get Report Description for a Specific bookmark_id}
\usage{
GetReportDescription(bookmark.id)
}
\arguments{
\item{bookmark.id}{}
}
\value{
List
}
\description{
Get report description for a specific bookmark_id
}
\details{
Requires a single bookmark_id value
}
\examples{
\dontrun{
reportdesc <- GetReportDescription("28473595")


}
}

