% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportSuiteGroups.R
\name{GetReportSuiteGroups}
\alias{GetReportSuiteGroups}
\title{Get Report Suite Groups for a specific report suite}
\usage{
GetReportSuiteGroups(reportsuite.id)
}
\arguments{
\item{reportsuite.id}{Report Suite ID}
}
\value{
Data frame
}
\description{
Retrieves a list of permission groups assigned to the specified report suite
}
\details{
Requires a single report suite id
}
\examples{
\dontrun{
rsg <- GetReportSuiteGroups("your-report-suite")


}
}
