\name{GetReportSuites}
\alias{GetReportSuites}
\title{Get Report Suites Associated with a Specific User/Company}
\usage{
GetReportSuites()
}
\value{
List of valid reportsuites
}
\description{
GetReportSuites
}
\details{
Gets all available report suites for the company.
}
\examples{
\dontrun{
reportsuites <- GetReportSuites()
}
}

