% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetScheduledSpike}
\alias{GetScheduledSpike}
\title{Get Permanent Traffic Setting for a Report Suite(s)}
\usage{
GetScheduledSpike(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get permanent traffic setting for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
spike <- GetScheduledSpike("your_report_suite")

spike2 <- GetScheduledSpike(report_suites$rsid)
}
}

