% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetSegments}
\alias{GetSegments}
\title{Get Segments Defined within a Report Suite}
\usage{
<<<<<<< HEAD
GetSegments(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get a data frame of segments for the specified report suites.
Useful to find segment IDs for use in reporting helper functions or JSON report definitions.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
segments <- GetSegments("your_report_suite")

segments2 <- GetSegments(report_suites$rsid)
}
}
=======
GetSegments(report_suites)
}
\arguments{
  \item{report_suites}{Character vector containing one or
  more valid Report Suite names}
}
\value{
Data Frame
}
\description{
Get Segments defined within a Report Suite. Segments
include Data Warehouse segments, pre-defined segments, and
user-defined segments
}
\details{
This function requires having a character vector with one
or more valid Report Suites specified.
}
\examples{
\dontrun{

    GetSegments("keystonejowanza")
    GetSegments(c("keystonejowanza", "keystonerandy", "keystonetraining"))
}
}
\keyword{segments}
>>>>>>> master

