% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSiteTitle.R
\name{GetSiteTitle}
\alias{GetSiteTitle}
\title{Get Site Title for a Report Suite(s)}
\usage{
GetSiteTitle(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get site title for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
sitetitle <- GetSiteTitle("your_report_suite")

sitetitle2 <- GetSiteTitle(report_suites$rsid)
}
}
