% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetTemplate.R
\name{GetTemplate}
\alias{GetTemplate}
\title{Get Template a Report Suite is Based On}
\usage{
GetTemplate(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get template a report suite is based on for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{

privacy <- GetPrivacySettings("your_report_suite")

privacy2 <- GetPrivacySettings(c("your_dev_suite", "your_prod_suite"))

}
}

