% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTransactionEnabled.R
\name{GetTransactionEnabled}
\alias{GetTransactionEnabled}
\title{Get Whether Transaction Storage for a Report Suite(s)}
\usage{
GetTransactionEnabled(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get whether transaction storage is enabled for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
trans <- GetTransactionEnabled("your_report_suite")

trans2 <- GetTransactionEnabled(report_suites$rsid)
}
}
