% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetUniqueVisitorVariable}
\alias{GetUniqueVisitorVariable}
\title{Get Whether Transaction Storage for a Report Suite(s)}
\usage{
GetUniqueVisitorVariable(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get whether transaction storage is enabled for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
uniq <- GetUniqueVisitorVariable("your_report_suite")

uniq2 <- GetUniqueVisitorVariable(report_suites$rsid)
}
}

