% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetUsageLog}
\alias{GetUsageLog}
<<<<<<< HEAD
\title{Get Admin Actions, Logins, and Reports Accessed}
\usage{
GetUsageLog(date.from = as.character(Sys.Date() - 1),
  date.to = as.character(Sys.Date()), localtime = FALSE)
}
\arguments{
\item{date.from}{Log start date (YYYY-MM-DD)}

\item{date.to}{Log end date (YYYY-MM-DD)}

\item{localtime}{Whether to change datetimes from UTC to local time}
}
\value{
data frame with: datetime, login, event_num, event_type, ip_address, report_suite, event_details
}
\description{
GetUsageLog
}
\details{
Gets the Adobe Analytics usage log for all users within the specified date range.
=======
\title{Admin Actions, Logins, and Reports Accessed}
\usage{
GetUsageLog(date_from = as.character(Sys.Date() - 1),
  date_to = as.character(Sys.Date()), localtime = FALSE)
}
\arguments{
  \item{date_from}{Optional: Start date of report, in
  "YYYY-MM-DD" format. If argument not set in function
  call, defaults to Sys.Date() - 1}

  \item{date_to}{Optional: End date of report, in
  "YYYY-MM-DD" format. If argument not set in function
  call, defaults to Sys.Date()}

  \item{localtime}{Optional. If TRUE, timestamp converted
  to local time. Otherwise, UNIX time in seconds.}
}
\value{
Data Frame
}
\description{
Creates data frame containing all Report Suite actions
(Admin, Logins, Reports Accessed) during a given time
period.
}
\details{
This report is administrative in nature, for users
interested in understanding how internal constituents are
using SiteCatalyst. This report is also a record of all
Admin changes to an account.
>>>>>>> master
}
\examples{
\dontrun{
usagelog <- GetUsageLog("2014-01-01","2014-01-31")
}
}
\seealso{
Other internal: \code{\link{ApiRequest}};
  \code{\link{BuildHeader}};
  \code{\link{BuildInnerBreakdownsRecursively}};
  \code{\link{GetEndpoint}};
  \code{\link{JsonQueueRealTimeReport}};
  \code{\link{JsonQueueReport}};
  \code{\link{ParseFallout}}; \code{\link{ParseOvertime}};
  \code{\link{ParsePathing}}; \code{\link{ParseRanked}};
  \code{\link{ParseTrended}}
}
\keyword{internal}

<<<<<<< HEAD
=======
    #Timestamp in local time
    GetUsageLog("2013-01-01", "2013-02-13", TRUE)

    #UNIX time in seconds for timestamp
    GetUsageLog("2013-01-01", "2013-02-13")
 }
}
\keyword{admin}

>>>>>>> master
