% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetVersionAccess}
\alias{GetVersionAccess}
\title{Get Products/Versions associated with a specific company}
\usage{
GetVersionAccess()
}
\value{
DataFrame
}
\description{
Get Products/Versions associated with a specific company
}
\details{
Returns a Data Frame of Adobe Analytics products
}
\examples{
\dontrun{
versions <- GetVersionAccess()
}
}

