% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVideoSettings.R
\name{GetVideoSettings}
\alias{GetVideoSettings}
\title{Get Video Settings for a Report Suite(s)}
\usage{
GetVideoSettings(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get video settings for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
vidsettings<- GetVideoSettings("your_report_suite")

vidsettings2 <- GetVideoSettings(report_suites$rsid)
}
}
