% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVirtualReportSuiteSettings.R
\name{GetVirtualReportSuiteSettings}
\alias{GetVirtualReportSuiteSettings}
\title{Get Virtual Report Suite Settings}
\usage{
GetVirtualReportSuiteSettings(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get Virtual Report Suite defintions.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
virtualsettings <- GetVirtualReportSuiteSettings("your_report_suite")

virtualsettings2 <- GetVirtualReportSuiteSettings(report_suites$rsid)

}
}
