% Generated by roxygen2 (4.0.1): do not edit by hand
\name{JsonQueueRealTimeReport}
\alias{JsonQueueRealTimeReport}
\title{Create Real-Time Report from JSON}
\usage{
JsonQueueRealTimeReport(report.description)
}
\arguments{
\item{report.description}{JSON report description}
}
\value{
Data frame
}
\description{
Generic interface to validate, queue and retrieve a realtime report from the API
}
\details{
This is a function for advanced users, for the case where a user feels
that submitting a JSON request would be easier than using one of the pre-defined
functions from RSiteCatalyst
}
\examples{
\dontrun{

custom_report <- JsonQueueRealTimeReport('valid Adobe Analytics API JSON string')

}
}

