% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseTrended.R
\name{ParseTrended}
\alias{ParseTrended}
\title{ParseTrended}
\usage{
ParseTrended(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses a trended report returned from the API
}
\seealso{
Other internal: \code{\link{GetEndpoint}},
  \code{\link{GetUsageLog}}, \code{\link{ParseDW}},
  \code{\link{ParseFallout}}, \code{\link{ParseOvertime}},
  \code{\link{ParsePathing}}, \code{\link{ParseRanked}},
  \code{\link{ParseSummary}}, \code{\link{call.Get_base}},
  \code{\link{collapse_simple_listcol}},
  \code{\link{collapse_simple_target}},
  \code{\link{split_segment_ret}}
}
\keyword{internal}
