\name{RSiteCatalyst}
\alias{RSiteCatalyst}
\docType{package}
\title{
Adobe Analytics (Omniture SiteCatalyst) Reporting API
}
\description{
This package contains an "analyst's toolbox" of functions for accessing the Adobe Analytics (Omniture SiteCatalyst) Reporting API. These functions allow the user to authenticate, get metadata about report suites (eVars, props, events, segments, etc.), and create reports using SiteCatalyst data. 

This package is not intended for SiteCatalyst Administration.
}
\details{
\tabular{ll}{
Package: \tab RSiteCatalyst\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-07-21\cr
License: \tab GPL-2\cr
}
}
\author{
Randy Zwitch & Jowanza Joseph

Maintainer: Randy Zwitch <rzwitch+rsitecatalyst@gmail.com>

}
\references{
Official Adobe Analytics (Omniture SiteCatalyst) API documentation: \cr \cr
https://developer.omniture.com/en_US/documentation
}

\keyword{ package }