<<<<<<< HEAD
% Generated by roxygen2 (4.0.1): do not edit by hand
=======
>>>>>>> master
\docType{package}
\name{RSiteCatalyst}
\alias{RSiteCatalyst}
\alias{RSiteCatalyst-package}
<<<<<<< HEAD
\title{R Client for Adobe Analytics API v1.4}
\description{
This package contains an "analyst's toolbox" of functions for accessing the Adobe Analytics Reporting API v1.4.
These functions allow the user to authenticate, get metadata about report suites (eVars, props, events, segments, etc.), and create reports using Adobe Analytics data.

This package is not intended for Adobe Analytics administration.
}
\details{
Package:  RSiteCatalyst \cr
Type:     Package \cr
Version:  1.4 \cr
Date:     2014-07-01 \cr
License:  MIT + file LICENSE \cr
}
\author{
Willem Paling, Randy Zwitch, Jowanza Joseph
}
\references{
Official Adobe Analytics API documentation:

https://marketing.adobe.com/developer/en_US/documentation

For support & bugs:

https://github.com/randyzwitch/RSiteCatalyst
}
=======
\title{Adobe Analytics Reporting API}
\description{
This package contains an "analyst's toolbox" of functions
for accessing the Adobe Analytics Reporting API v1.3. These
functions allow the user to authenticate, get metadata
about report suites (eVars, props, events, segments, etc.),
and create reports using Adobe Analytics data. This package
does not support the v1.4 API that uses OAuth for
authentication.
}
\details{
This package is not intended for Adobe Analytics
Administration.

\tabular{ll}{ Package: \tab RSiteCatalyst\cr Type: \tab
Package\cr Version: \tab 1.3.3\cr Date: \tab 2014-04-06\cr
License: \tab GPL-2\cr }
}
\author{
Randy Zwitch & Jowanza Joseph

Maintainer: Randy Zwitch <rzwitch+rsitecatalyst@gmail.com>
}
\references{
Official Adobe Analytics API documentation: \cr \cr
https://developer.omniture.com/en_US/documentation

For support & bugs, please use GitHub repository: \cr \cr
https://github.com/randyzwitch/RSiteCatalyst
}
\keyword{package}
>>>>>>> master

