% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateReport.R
\name{ValidateReport}
\alias{ValidateReport}
\title{ValidateReport}
\usage{
ValidateReport(report.description, interval.seconds = 0, max.attempts = 1)
}
\arguments{
\item{report.description}{JSON report description}

\item{interval.seconds}{Time to wait between attempts to validate the report (defaults to 2 seconds)}

\item{max.attempts}{Max number of attempts to make to validate the report (defaults to 1)}
}
\value{
TRUE/FALSE depending on whether the report is valid or not
}
\description{
Internal function - Calls the API and attempts to validate a report description.
}
\keyword{internal}
