% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewProcessingRules.R
\name{ViewProcessingRules}
\alias{ViewProcessingRules}
\title{View Processing Rules}
\usage{
ViewProcessingRules(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get Processing Rules with title and actions.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{

 pr <- ViewProcessingRules("your-report-suite")
 pr <- ViewProcessingRules(c("your-report-suite", "your-report-suite2"))
 
}
}
