% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiRequest.R
\name{ApiRequest}
\alias{ApiRequest}
\title{ApiRequest}
\usage{
ApiRequest(body = "", func.name = "", interval.seconds = 2,
  max.attempts = 1, print.attempts = FALSE, skip.queue = FALSE,
  format = "json")
}
\arguments{
\item{func.name}{the name of the Adobe Analytics API function that we are calling}

\item{interval.seconds}{Time to wait between request attempts (defaults to 2 seconds)}

\item{max.attempts}{Max number of attempts to make the request (defaults to 1, this is only increased for GetReport)}

\item{print.attempts}{if set to TRUE, this will print attempt numbers to the console}

\item{skip.queue}{If set to TRUE, return content instead of parsing via jsonlite}

\item{report.description}{JSON report description}
}
\value{
json
}
\description{
Internal function - Calls the API and gets response for specified params
}
\keyword{internal}
