% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BuildHeader}
\alias{BuildHeader}
\title{BuildHeader}
\usage{
BuildHeader()
}
\value{
Header string for the old API auth method.
}
\description{
Internal function - Build Header for REST API call. This is deprecated, we should now be using OAUTH.
}
\seealso{
Other internal: \code{\link{ApiRequest}};
  \code{\link{BuildInnerBreakdownsRecursively}};
  \code{\link{GetEndpoint}}; \code{\link{GetUsageLog}};
  \code{\link{JsonQueueRealTimeReport}};
  \code{\link{JsonQueueReport}};
  \code{\link{ParseFallout}}; \code{\link{ParseOvertime}};
  \code{\link{ParsePathing}}; \code{\link{ParseRanked}};
  \code{\link{ParseTrended}}
}
\keyword{internal}

