% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHeader.R
\name{BuildHeader}
\alias{BuildHeader}
\title{BuildHeader}
\usage{
BuildHeader()
}
\value{
Header string for the old API auth method.
}
\description{
Internal function - Build Header for REST API call. This is deprecated, we should now be using OAUTH.
}
\keyword{internal}
