% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CancelReport}
\alias{CancelReport}
\title{Cancel a Report in the Report Queue}
\usage{
<<<<<<< HEAD
CancelReport(report.id)
}
\arguments{
\item{report.id}{Id of the report that you want to cancel}
}
\value{
Console message
}
\description{
Cancels a report in the Report Queue
}
\details{
Returns either a console message that no reports are queued
or the reportID number that was cancelled
}
\examples{
\dontrun{
CancelReport('12345678')
}
}
=======
CancelReport(reportID)
}
\arguments{
  \item{reportID}{Report ID obtained from GetReportQueue}
}
\value{
Character string or List
}
\description{
Cancels a report in the Report Queue
}
\details{
Returns either a console message that no reports are queued
or a list of reportID numbers
}
\examples{
\dontrun{

    CancelReport(12345678)

    #Example Result
    "Report 12345678 has been cancelled"

          }
}
\seealso{
\code{\link{GetReportQueue}}
}
\keyword{cancel}
\keyword{report}
>>>>>>> master

