% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancelReport.R
\name{CancelReport}
\alias{CancelReport}
\title{Cancel a Report in the Report Queue}
\usage{
CancelReport(report.id)
}
\arguments{
\item{report.id}{Id of the report that you want to cancel}
}
\value{
Console message
}
\description{
Cancels a report in the Report Queue
}
\details{
Returns either a console message that no reports are queued 
or the reportID number that was cancelled
}
\examples{
\dontrun{
CancelReport('12345678')
}
}
