% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCustomCalendar.R
\name{GetCustomCalendar}
\alias{GetCustomCalendar}
\title{Get Custom Calendar for a Report Suite(s)}
\usage{
GetCustomCalendar(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get custom calendar for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
cal <- GetCustomCalendar("your_report_suite")

cal2 <- GetCustomCalendar(report_suites$rsid)
}
}
