% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDashboards.R
\name{GetDashboards}
\alias{GetDashboards}
\title{Get Defined Dashboards}
\usage{
GetDashboards(dashboard.limit = "", dashboard.offset = "")
}
\arguments{
\item{dashboard.limit}{Limit number of dashboards returned}

\item{dashboard.offset}{Offset of dashboards (i.e. start with other than first dashboard)}
}
\value{
List
}
\description{
Get defined dashboards
}
\details{
This function's arguments are both optional
}
\examples{
\dontrun{
dash<- GetDashboards()

dash2 <- GetBookmarks('5', '1')
}
}
