% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataWarehouseDisplay.R
\name{GetDataWarehouseDisplay}
\alias{GetDataWarehouseDisplay}
\title{Get Whether Data Warehouse is Enabled for a Report Suite(s)}
\usage{
GetDataWarehouseDisplay(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get whether Data Warehouse is enabled for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
url <- GetDataWarehouseDisplay("your_report_suite")

url2 <- GetDataWarehouseDisplay(report_suites$rsid)
}
}
