% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEcommerce.R
\name{GetEcommerce}
\alias{GetEcommerce}
\title{Get the Commerce Level for a Report Suite(s)}
\usage{
GetEcommerce(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get the commerce level for each of the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
ecom <- GetEcommerce("your_report_suite")

ecom2 <- GetEcommerce(report_suites$rsid)
}
}
