% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEVars.R
\name{GetEvars}
\alias{GetEvars}
\title{Get Commerce Variables (eVars) Associated with a Report Suite}
\usage{
GetEvars(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get Commerce Variables (eVars) Associated with a Report Suite.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
evars <- GetEvars("your_report_suite")

evars2 <- GetEvars(report_suites$rsid)

}
}
