% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarketingChannelExpiration.R
\name{GetMarketingChannelExpiration}
\alias{GetMarketingChannelExpiration}
\title{Get Marketing Channel Expiration for a Report Suite(s)}
\usage{
GetMarketingChannelExpiration(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get marketing channel expiration for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
expire <- GetMarketingChannelExpiration("your_report_suite")

expire2 <- GetMarketingChannelExpiration(report_suites$rsid)
}
}
