% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPaidSearchDetection.R
\name{GetPaidSearchDetection}
\alias{GetPaidSearchDetection}
\title{Get Paid Search Detection Parameters for a Report Suite(s)}
\usage{
GetPaidSearchDetection(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get paid search detection parameters for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
paidsearch <- GetPaidSearchDetection("your_report_suite")

paidsearch2 <- GetPaidSearchDetection(report_suites$rsid)
}
}
