% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProps.R
\name{GetProps}
\alias{GetProps}
\title{Get Traffic Variables (props) Associated with a Report Suite}
\usage{
GetProps(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get Traffic Variables (props) Associated with a Report Suite(s).
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
props <- GetProps("your_report_suite")

props2 <- GetProps(report_suites$rsid)
}
}
