% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQueue.R
\name{GetQueue}
\alias{GetQueue}
\title{Get Number/ID of Reports in Queue}
\usage{
GetQueue()
}
\value{
Console message and/or list
}
\description{
Requests the number of reports in the Report Queue, as well as the Report ID.
}
\details{
Returns either a message to the console that no reports are in the Queue or a list with the Report ID's.
}
\examples{
\dontrun{
queue <- GetQueue()
}
}
