% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetRealTimeReport}
\alias{GetRealTimeReport}
\title{Get Real-Time report}
\usage{
GetRealTimeReport(reportsuite.ids, metrics, elements = c(),
  date.granularity = 5, date.from = "1 hour ago", date.to = "now",
  sort.algorithm = "mostpopular", floor.sensitivity = 0.25,
  first.rank.period = 0, algorithm.argument = "linear",
  everything.else = TRUE)
}
\arguments{
\item{reportsuite.ids}{Report Suite}

\item{metrics}{Report metric}

\item{elements}{Report breakdowns}

\item{date.granularity}{Report Granularity. Defaults to 5 minutes}

\item{date.from}{Report starting time. Defaults to "1 hour ago"}

\item{date.to}{Report end time. Defaults to "now"}

\item{sort.algorithm}{Sorting algorithm. Defaults to "mostpopular"}

\item{floor.sensitivity}{Floor sensitivity. Defaults to .25}

\item{first.rank.period}{First Ranking Period. Defaults to 0}

\item{algorithm.argument}{Ranking algorithm. Defaults to "linear"}

\item{everything.else}{Provide counts for elements not returned as 'top'}
}
\value{
Data frame
}
\description{
Fill in later
}
\details{
Fill in later
}
\examples{
\dontrun{

custom_report <- GetRealTimeReport('')

}
}

