<<<<<<< HEAD
% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetRealTimeReport}
\alias{GetRealTimeReport}
\title{Get Real-Time report}
\usage{
GetRealTimeReport(reportsuite.ids, metrics, elements = c(),
  date.granularity = 5, date.from = "1 hour ago", date.to = "now",
  sort.algorithm = "mostpopular", floor.sensitivity = 0.25,
  first.rank.period = 0, algorithm.argument = "linear",
  everything.else = TRUE)
}
\arguments{
\item{reportsuite.ids}{Report Suite}

\item{metrics}{Report metric}

\item{elements}{Report breakdowns}

\item{date.granularity}{Report Granularity. Defaults to 5 minutes}

\item{date.from}{Report starting time. Defaults to "1 hour ago"}

\item{date.to}{Report end time. Defaults to "now"}

\item{sort.algorithm}{Sorting algorithm. Defaults to "mostpopular"}

\item{floor.sensitivity}{Floor sensitivity. Defaults to .25}

\item{first.rank.period}{First Ranking Period. Defaults to 0}

\item{algorithm.argument}{Ranking algorithm. Defaults to "linear"}

\item{everything.else}{Provide counts for elements not returned as 'top'}
}
\value{
Data frame
}
\description{
Function to access the Adobe Analytics Real-Time API v1.4. This API
provides the ability for reporting up to the most recent minute. This API is best
used at 15-30 second intervals (or longer).
}
\details{
The Real-Time API uses a concept of "relative dates". To get a feeling for
what's possible for submitting to date.from and date.to parameters, see link at:

http://php.net/manual/en/datetime.formats.relative.php
=======
\name{GetRealTimeReport}
\alias{GetRealTimeReport}
\title{Get Real-Time Report}
\usage{
GetRealTimeReport(report_suite, metrics, elements = c(), periodMinutes = 1,
  periodCount = 15, periodOffset = 0, algorithm = "most popular",
  algorithmArgument = "linear", firstRankPeriod = "0",
  floorSensitivity = 0.25)
}
\arguments{
  \item{report_suite}{Report Suite ID}

  \item{metrics}{Metric to include in the Realtime report.
  The metric must be configured previously using
  SaveRealTimeConfiguration}

  \item{elements}{A list of elements that breaks down
  (organizes) the metrics data in the report. Each element
  must be configured previously using
  SaveRealTimeConfiguration.

  If no elements are provided, an overtime report is
  generated for the provided metric.}

  \item{periodMinutes}{Number of minutes for one period.
  Default is 1.}

  \item{periodCount}{The number of periods of data to
  return. Default is 15.}

  \item{periodOffset}{Number of minutes before the current
  minute to run the report. If set to 10, the most recent
  result will be from ten minutes prior to the request.
  Default is 0.}

  \item{algorithm}{Type of dimensions to return, one of the
  following three values: 'gainers', 'losers', 'most
  popular'. Default is 'most popular'.}

  \item{algorithmArgument}{Specifies how to order the
  values for Most Popular, Gainers or Losers. Specify
  either percent, or linear. Default is linear.}

  \item{firstRankPeriod}{Computes the ranking of elements
  by considering the element's counts from the
  firstRankPeriod to the final period. With this argument
  you can rank from the first period (0) to periodCount - 1
  (most popular) or periodCount - 3 (gainers/losers) or
  anywhere in between. The firstRankPeriod is 0 based.}

  \item{floorSensitivity}{A factor between 0 and 1 that is
  used to cut off low-count items from percentage ranking.
  Relative only to gainers/losers by percent. Default is
  .25.}
}
\value{
Data Frame
}
\description{
Get real-time report
}
\details{
GetRealTimeReport returns a Data Frame with the results
from the real-time report specified. To find out valid
values for the inputs of this function, run
GetRealTimeConfiguration to get the current report setup.
To change the real-time report setup, use
SaveRealTimeConfiguration.

Currently, RSiteCatalyst only supports up to one element in
a realtime report. This limitation will be removed in a
future release.
>>>>>>> master
}
\examples{
\dontrun{

<<<<<<< HEAD
custom_report <- GetRealTimeReport('')

}
}
=======
  #Minimum example - Equivalent to Overtime report
  GetRealTimeReport("keystonerandy", "instances")

   }
}
\references{
See documentation for in-depth documentation:

https://developer.omniture.com/en_US/documentation/sitecatalyst-reporting/r-realtimereportdescription
}
\seealso{
\code{\link{GetRealTimeConfiguration}}\cr
\code{\link{SaveRealTimeConfiguration}}\cr
}
\keyword{GetRealTimeConfiguration}
>>>>>>> master

