% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportDescription.R
\name{GetReportDescription}
\alias{GetReportDescription}
\title{Get Report Description for a Specific bookmark_id}
\usage{
GetReportDescription(bookmark.id)
}
\arguments{
\item{bookmark.id}{Bookmark ID}
}
\value{
List
}
\description{
Get report description for a specific bookmark_id
}
\details{
Requires a single bookmark_id value, obtained from GetBookmarks()
}
\examples{
\dontrun{
reportdesc <- GetReportDescription("28473595")


}
}
