\name{GetReportSuites}
\alias{GetReportSuites}
\title{Get Report Suites Associated with a Specific User/Company}
\description{
    Get Report Suites Associated with a Specific User/Company
}
\usage{
    GetReportSuites()
}
\arguments{
    None
}
\details{
    Returns a data frame containing the Report Suite ID and Site Title
}
\value{
    Data Frame
}

\examples{
\dontrun{    
    
    GetReportSuites()
    }
    
    
    
}
\keyword{suite}