% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportSuites.R
\name{GetReportSuites}
\alias{GetReportSuites}
\title{Get Report Suites Associated with a Specific User/Company}
\usage{
GetReportSuites()
}
\value{
Data frame
}
\description{
Get Report Suites Associated with a Specific User/Company
}
\details{
Returns a data frame containing the Report Suite ID and Site Title
}
\examples{
\dontrun{
report_suites <- GetReportSuites()
}
}
