% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetScheduledSpike.R
\name{GetScheduledSpike}
\alias{GetScheduledSpike}
\title{Get Scheduled Traffic Spike Setting for a Report Suite(s)}
\usage{
GetScheduledSpike(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get scheduled traffic spikes for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
spike <- GetScheduledSpike("your_report_suite")

spike2 <- GetScheduledSpike(report_suites$rsid)
}
}
