% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSegments.R
\name{GetSegments}
\alias{GetSegments}
\title{Get Segments Defined within a Report Suite}
\usage{
GetSegments(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get a data frame of segments for the specified report suites.
Useful to find segment IDs for use in reporting helper functions or JSON report definitions.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
segments <- GetSegments("your_report_suite")

segments2 <- GetSegments(report_suites$rsid)
}
}
