% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTimeZone.R
\name{GetTimeZone}
\alias{GetTimeZone}
\title{Get Time Zone for a Report Suite(s)}
\usage{
GetTimeZone(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get Time Zone for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
timezone <- GetTimeZone("your_report_suite")

timezone2 <- GetTimeZone(report_suites$rsid)
}
}
