% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTrackingServer.R
\name{GetTrackingServer}
\alias{GetTrackingServer}
\title{Get Tracking Server Associated with a Namespace (Company)}
\usage{
GetTrackingServer(reportsuite.id)
}
\arguments{
\item{reportsuite.id}{report suite id}
}
\value{
Data frame
}
\description{
Get tracking server associated with a namespace (company).
}
\details{
This function requires having a character string with a valid Report
Suite specified. You can specify any report suite you want, as all report 
suites have same tracking server.
}
\examples{
\dontrun{
ts <- GetTrackingServer("your_report_suite")


}
}
