% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUniqueVisitorVariable.R
\name{GetUniqueVisitorVariable}
\alias{GetUniqueVisitorVariable}
\title{Get Whether Unique Visitor Variable Enabled for a Report Suite(s)}
\usage{
GetUniqueVisitorVariable(reportsuite.ids)
}
\arguments{
\item{reportsuite.ids}{Report suite id (or list of report suite ids)}
}
\value{
Data frame
}
\description{
Get whether unique visitor variable is enabled for the specified report suites.
}
\details{
This function requires having a character vector with one or more valid Report Suites specified.
}
\examples{
\dontrun{
uniq <- GetUniqueVisitorVariable("your_report_suite")

uniq2 <- GetUniqueVisitorVariable(report_suites$rsid)
}
}
