% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVersionAccess.R
\name{GetVersionAccess}
\alias{GetVersionAccess}
\title{Get Products/Versions associated with a specific company}
\usage{
GetVersionAccess()
}
\value{
DataFrame
}
\description{
Get Products/Versions associated with a specific company
}
\details{
Returns a Data Frame of Adobe Analytics products
}
\examples{
\dontrun{
versions <- GetVersionAccess()
}
}
