\name{JsonQueueRealTimeReport}
\alias{JsonQueueRealTimeReport}
\title{JsonQueueRealTimeReport}
\usage{
JsonQueueRealTimeReport(report.description)
}
\arguments{
  \item{report.description}{JSON report description}
}
\value{
Data frame
}
\description{
Generic interface to validate, queue and retrieve a
realtime report from the API
}

