% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JsonQueueRealTimeReport}
\alias{JsonQueueRealTimeReport}
\title{Create Real-Time Report from JSON}
\usage{
JsonQueueRealTimeReport(report.description)
}
\arguments{
\item{report.description}{JSON report description}
}
\value{
Data frame
}
\description{
Generic interface to validate, queue and retrieve a realtime report from the API
}
\details{
Used in GetRealTimeReport
}
\examples{
\dontrun{

custom_report <- JsonQueueRealTimeReport('valid Adobe Analytics API JSON string')

}
}
\seealso{
Other internal: \code{\link{ApiRequest}};
  \code{\link{BuildHeader}};
  \code{\link{BuildInnerBreakdownsRecursively}};
  \code{\link{GetEndpoint}}; \code{\link{GetUsageLog}};
  \code{\link{ParseFallout}}; \code{\link{ParseOvertime}};
  \code{\link{ParsePathing}}; \code{\link{ParseRanked}};
  \code{\link{ParseTrended}}
}
\keyword{internal}

