\name{JsonQueueReport}
\alias{JsonQueueReport}
\title{JsonQueueReport}
\usage{
JsonQueueReport(report.description)
}
\arguments{
  \item{report.description}{JSON report description}
}
\value{
Formatted data frame
}
\description{
Generic interface to validate, queue and retrieve a report
from the API
}

