% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ParseFallout}
\alias{ParseFallout}
\title{ParseFallout}
\usage{
ParseFallout(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Data frame
}
\description{
Internal Function - Parses a fallout report returned from the API
}
\seealso{
Other internal: \code{\link{ApiRequest}};
  \code{\link{BuildHeader}};
  \code{\link{BuildInnerBreakdownsRecursively}};
  \code{\link{GetEndpoint}}; \code{\link{GetUsageLog}};
  \code{\link{JsonQueueRealTimeReport}};
  \code{\link{ParseOvertime}}; \code{\link{ParsePathing}};
  \code{\link{ParseRanked}}; \code{\link{ParseTrended}}
}
\keyword{internal}

