% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseFallout.R
\name{ParseFallout}
\alias{ParseFallout}
\title{ParseFallout}
\usage{
ParseFallout(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Data frame
}
\description{
Internal Function - Parses a fallout report returned from the API
}
\seealso{
Other internal: \code{\link{GetEndpoint}},
  \code{\link{GetUsageLog}}, \code{\link{ParseDW}},
  \code{\link{ParseOvertime}}, \code{\link{ParsePathing}},
  \code{\link{ParseRanked}}, \code{\link{ParseSummary}},
  \code{\link{ParseTrended}}
}
\keyword{internal}
