% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ParseOvertime}
\alias{ParseOvertime}
\title{ParseOvertime}
\usage{
ParseOvertime(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses an overtime report returned from the API
}
\seealso{
Other internal: \code{\link{ApiRequest}};
  \code{\link{BuildHeader}};
  \code{\link{BuildInnerBreakdownsRecursively}};
  \code{\link{GetEndpoint}}; \code{\link{GetUsageLog}};
  \code{\link{JsonQueueRealTimeReport}};
  \code{\link{ParseFallout}}; \code{\link{ParsePathing}};
  \code{\link{ParseRanked}}; \code{\link{ParseTrended}}
}
\keyword{internal}

