% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ParseOvertime.R
\name{ParseOvertime}
\alias{ParseOvertime}
\title{ParseOvertime}
\usage{
ParseOvertime(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses an overtime report returned from the API
}
\seealso{
Other internal: \code{\link{GetEndpoint}};
  \code{\link{GetUsageLog}}; \code{\link{ParseFallout}};
  \code{\link{ParsePathing}}; \code{\link{ParseRanked}};
  \code{\link{ParseSummary}}; \code{\link{ParseTrended}}
}
\keyword{internal}

