% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParsePathing.R
\name{ParsePathing}
\alias{ParsePathing}
\title{ParsePathing}
\usage{
ParsePathing(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses a pathing report returned from the API
}
\seealso{
Other internal: \code{\link{GetEndpoint}},
  \code{\link{GetUsageLog}}, \code{\link{ParseDW}},
  \code{\link{ParseFallout}}, \code{\link{ParseOvertime}},
  \code{\link{ParseRanked}}, \code{\link{ParseSummary}},
  \code{\link{ParseTrended}}
}
\keyword{internal}
