% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ParseRanked.R
\name{ParseRanked}
\alias{ParseRanked}
\title{ParseRanked}
\usage{
ParseRanked(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses a ranked report returned from the API
}
\seealso{
Other internal: \code{\link{GetEndpoint}};
  \code{\link{GetUsageLog}}; \code{\link{ParseFallout}};
  \code{\link{ParseOvertime}}; \code{\link{ParsePathing}};
  \code{\link{ParseTrended}}
}
\keyword{internal}

