% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseSummary.R
\name{ParseSummary}
\alias{ParseSummary}
\title{ParseSummary}
\usage{
ParseSummary(report.data)
}
\arguments{
\item{report.data}{jsonlite formatted data frame of report data returned from the API}
}
\value{
Data frame
}
\description{
Internal Function - Parses a summary report returned from the API
}
\seealso{
Other internal: \code{\link{GetEndpoint}},
  \code{\link{GetUsageLog}}, \code{\link{ParseDW}},
  \code{\link{ParseFallout}}, \code{\link{ParseOvertime}},
  \code{\link{ParsePathing}}, \code{\link{ParseRanked}},
  \code{\link{ParseTrended}}
}
\keyword{internal}
