\name{RSiteCatalyst}
\alias{RSiteCatalyst}
\docType{package}
\title{
Adobe Analytics Reporting API
}
\description{
This package contains an "analyst's toolbox" of functions for accessing the Adobe Analytics Reporting API v1.3. These functions allow the user to authenticate, get metadata about report suites (eVars, props, events, segments, etc.), and create reports using Adobe Analytics data. This package does not support the v1.4 API that uses OAuth for authentication. 

This package is not intended for Adobe Analytics Administration.
}
\details{
\tabular{ll}{
Package: \tab RSiteCatalyst\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2014-01-28\cr
License: \tab GPL-2\cr
}
}
\author{
Randy Zwitch & Jowanza Joseph

Maintainer: Randy Zwitch <rzwitch+rsitecatalyst@gmail.com>

}
\references{
Official Adobe Analytics API documentation: \cr \cr
https://developer.omniture.com/en_US/documentation

For support & bugs, please use GitHub repository: \cr \cr
https://github.com/randyzwitch/RSiteCatalyst
}

\keyword{ package }