\docType{package}
\name{RSiteCatalyst}
\alias{RSiteCatalyst}
\alias{RSiteCatalyst-package}
\title{R Client for Adobe Analytics v1.4 API}
\description{
R Client for Adobe Analytics v1.4 API
}
\details{
R client library for the Adobe Analytics 1.4 API, currently
in beta, but due out in Feb 2014.
}
\author{
Willem Paling, Randy Zwitch, Jowanza Joseph
}

