% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RSiteCatalyst.R
\docType{package}
\name{RSiteCatalyst}
\alias{RSiteCatalyst}
\alias{RSiteCatalyst-package}
\title{R Client for Adobe Analytics API V1.4}
\description{
This package contains an "analyst's toolbox" of functions for accessing the Adobe Analytics Reporting API v1.4.
These functions allow the user to authenticate, get metadata about report suites (eVars, props, events, segments, etc.), and create reports using Adobe Analytics data.

This package is not intended for Adobe Analytics administration.
}
\details{
Package:  RSiteCatalyst \cr
Type:     Package \cr
Version:  1.4.6 \cr
Date:     2015-09-18 \cr
License:  MIT + file LICENSE \cr
}
\author{
Willem Paling, Randy Zwitch, Jowanza Joseph
}
\references{
Official Adobe Analytics API documentation:

https://marketing.adobe.com/developer/en_US/documentation

For support & bugs:

https://github.com/randyzwitch/RSiteCatalyst
}

