% Generated by roxygen2 (4.0.0): do not edit by hand
\name{SaveRealTimeSettings}
\alias{SaveRealTimeSettings}
\title{Save Configuration for Real-Time Report}
\usage{
SaveRealTimeSettings(reportsuite.ids = "", report1 = list(),
  report2 = list(), report3 = list())
}
\arguments{
\item{reportsuite.ids}{Report Suite ID}

\item{report1}{Real Time Report 1}

\item{report2}{Real Time Report 2}

\item{report3}{Real Time Report 3}
}
\value{
Message returned to console
}
\description{
Sets the metrics and elements (dimensions) on which you want real time
reports enabled via list objects created by BuildRealTimeReportStructure.
Realtime configuration changes take 15 minutes to be reflected in reports.
}
\details{
SaveRealTimeSettings should be called each time you want to modify the
structure of your real-time reports. If you are unsure of your current setup
of your real-time reports, use GetRealTimeSettings to find out your
current setup.

Changes can take up to 15 minutes to be reflected.
}
\examples{
\dontrun{

   }
}
\seealso{
\code{\link{GetRealTimeSettings}} \cr
}
\keyword{SaveRealTimeSettings}

