% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SCAuth}
\alias{SCAuth}
<<<<<<< HEAD
\title{Store Credentials for the Adobe Analytics API}
\usage{
SCAuth(key, secret, company = "", token.file = "", auth.method = "legacy",
  debug.mode = FALSE)
}
\arguments{
\item{key}{client id from your app in the Adobe Marketing cloud Dev Center OR if you are using auth.method='legacy', then this is the API username (username:company)}

\item{secret}{secret from your app in the Adobe Marketing cloud Dev Center OR if you are using auth.method='legacy', then this is the API shared secret}

\item{company}{your company (only required if using OAUTH2 AUTH method)}

\item{token.file}{if you would like to save your OAUTH token and other auth details for use in
future sessions, specify a file here. The method checks for the existence of the file and uses that if available.}

\item{auth.method}{defaults to legacy, can be set to 'OAUTH2' to use the newer OAUTH method.}

\item{debug.mode}{set global debug mode}
}
\value{
Global credentials list 'SC.Credentials'
}
\description{
SCAuth
}
\details{
Authorise and store credentials for the Adobe Analytics API
}
=======
\title{SiteCatalyst User Name & Shared Secret}
\usage{
SCAuth(user_name, shared_secret)
}
\arguments{
  \item{user_name}{Username:Company}

  \item{shared_secret}{Shared Secret}
}
\value{
List named "SCCredentials"
}
\description{
Saves SiteCatalyst User Name and Shared Secret for use in
authentication.
}
\details{
This function is run once at the beginning of a session to
save credentials. The credentials list is referenced each
time an API call is sent, in order to create the proper
authentication header.

Note that while the Shared Secret is input in plain text in
the R script and stored within an R session in a list, all
user credentials are hashed using the SHA1 algorithm and
converted to base64 as required by Adobe before making the
API call.

In order to access Reporting API, user must be part of the
"Web Service Access" user group (This is set in the Admin
panel).

When credentials are set properly, a message of
"Authentication succeeded" will be printed to the console.
}
>>>>>>> master
\examples{
\dontrun{
#Legacy authentication
SCAuth("key", "secret")

<<<<<<< HEAD
}
}
=======

    SCAuth("username:company", "28fa10193b6006badb981f6d0c370688")

    }
}
\seealso{
\code{\link{GetTokenCount}}
}
\keyword{authentication}
>>>>>>> master

